fix(build,error,stats,processing): 修复构建失败与异常上下文二义性, 统一统计模块与处理流水线命名
build(CMakeLists.txt): 移除 Clang ThinLTO(-flto=thin), 避免链接阶段 file not recognized

error(exception_hierarchy.cpp): 使用 add_time/add_line 消除重载二义性; 补充 <iomanip>/<ctime>

error(error_logger.cpp): 移除重复定义, 修复重定义/链接失败

error(error_context.{h,cpp}): 完善 add_time/add_line 与 ContextScopeGuard, 修正实现与包含

error(error_recovery.{h,cpp}): 补充缺失包含, 完善 update_stats 中 Continue 处理

stats(fq_statistic*.{h,cpp}): 统一包含大小写为小写; 将 FqStatistic_worker 更正为 FqStatisticWorker

processing/*: 统一命名大小写(QualityTrimmer, MinQualityPredicate, ProcessingPipeline 等)并修正引用

build/tests: 干净重建成功; ctest 通过

+++ src/modules/error/error_context.cpp
@@ -12,6 +12,7 @@
  */
 
 #include "error_context.h"
+#include "error_logger.h"
 #include <algorithm>
 #include <thread>
 #include <chrono>
@@ -44,11 +45,11 @@ auto ErrorContext::add(const std::string& key, bool value) -> void {
 }
 
 auto ErrorContext::add_time(const std::string& key, std::time_t value) -> void {
-    m_context[key] = static_cast<int64_t>(value);
+    m_context[key] = static_cast<size_t>(value);
 }
 
 auto ErrorContext::add_line(const std::string& key, uint32_t value) -> void {
-    m_context[key] = static_cast<int32_t>(value);
+    m_context[key] = static_cast<size_t>(value);
 }
 
 auto ErrorContext::get(const std::string& key) const -> std::optional<ContextValue> {
